requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-filter==23.3
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1
celery==5.3.4
psycopg2-binary==2.9.7
Pillow==10.1.0
python-magic==0.4.27
django-extensions==3.2.3
django-redis==5.4.0
python-decouple==3.8
whitenoise==6.6.0
gunicorn==21.2.0
daphne==4.0.0
python-multipart==0.0.6
cryptography==41.0.7
PyJWT==2.8.0
pdf2image>=1.16.0